Move class SentryConfig to marquez.tracing pkg
